Autonomous_Research_Agent/requirements.txt
streamlit==1.36.0
google-generativeai==0.7.2
duckduckgo-search>=6.2.1
newspaper3k==0.2.8
requests>=2.31.0
python-docx>=1.1.0
